Scripts for spot detection
For this assignment you must implement and evaluate two methods for spot
detection: thresholding and the Watershed algorithm. Notice that neither
of these methods can actually give you the location of the spots, so you
will have to find a way to label pixels according to which spot they
belong to and then compute the center of each group of pixels. You are
allowed to use a library for this part of the assignment (see this and
this). You should use the test image "spotdetection_testgel.gif" for
testing the algorithms, but you are free to use additional gel images
too.